fix (De)Serialization for trait objects
`typetag` crate does not work for target `wasm`, so I have to
implement  my own way of de/serializing `dyn Trait`.
Another problem is that Type `Obj` (`Rc<RefCell<_>>`) is used
to reference the same object from multiple locations, so we have
a major issue for deserialization (not fixed by this commit)!